agent/requirements.txt
livekit-agents>=1.3.0
livekit-plugins-google>=0.8.0
livekit-plugins-silero
python-dotenv